Break out gui.rs
